modify translate to be an action; mock up quick action executer function
